requirements.txt
attrs==25.4.0
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
cssselect==1.3.0
defusedxml==0.7.1
dnspython==2.8.0
filelock==3.20.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
itemadapter==0.12.2
itemloaders==1.3.2
jmespath==1.0.1
lxml==6.0.2
packaging==25.0
parsel==1.10.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDispatcher==2.0.7
pymongo==4.15.3
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
queuelib==1.8.0
requests==2.32.5
requests-file==3.0.1
Scrapy==2.13.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
tldextract==5.3.0
Twisted==25.5.0
typing_extensions==4.15.0
urllib3==2.5.0
w3lib==2.3.1
wheel==0.45.1
zope.interface==8.0.1
